update environment validation webhook
This will no longer require that GitRepositoryUpdates use exactly
one of Kustomize, Helm, or Bookkeeper. Now it will require AT MOST
one. The point of this change is to accommodate use cases where
the ONLY desired behavior is taking commits from the head of some
subscribed-to branch and writing them to the head of an
environment-specific branch.

Signed-off-by: Kent <kent.rancourt@gmail.com>