fix(cli): handle CLI config file not found error
Fixes https://github.com/akuity/kargo/issues/1341

Currently, the CLI entrypoint and some commands assumes that the config
file exists in the home directory. So all commands and config-related
commands (e.g. `config set`, `config unset`) are failing if the config
file does not exist.

This commit fixes this issue by returning an default config if the
config file does not exists in the CLI entrypoint and creates the
config file if necessary in config-related commands.

Signed-off-by: Sunghoon Kang <hoon@akuity.io>