feat(creds)!: refuse lookup for HTTP Git endpoint
In preparation of allowing insecure (non-TLS) HTTP endpoints for
Git URLs. Which we do want to allow to connect to  e.g. local
(development) servers, but without ever passing credentials to them.

To not introduce breaking changes for HTTP Helm repositories which
at present appear to allow passing basic auth credentials over
an insecure wire, we specifically check for the type to equal Git
before bailing. If we decide this is actually undesired behavior,
it can easily be removed.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>