fix(devx): mount permission issues on Linux
This addresses an issue observed on Linux systems where files written
to mount paths would end up being owned by root, or files could not
be read because they were _not_ owned by the container's user.

It is achieved by creating a user within the development image with
the same UID and GID as the caller on the host, while ensuring mount
target paths do exist for this user and group before the container
starts.

This effectively makes the `hack-build-cli` target work while also
delivering smoother operations with `hack-codegen`.

In the future, we should likely rethink the way we work with tooling
dependencies and their versions. As this problem was harder to tackle
due to the `codegen` target performing system-global installations.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>