fix(controller): do not return non-0 result and err
This avoids the warning from being logged, while also moving towards
more "controller-runtime best practices" as documented in #1479
(albeit not entirely related).

In general, I am of opinion that construction of the `ctrl.Result`
is best done at return time to provide a clear view on what the
expected requeue behavior would be without having to mentally
maintain a list of conditionals which could have affected the
returned result.

In case the preference would still be to apply any defaults,
I think this would be better done by defining a `defer` which
(conditionally) modifies the `res` and/or `err` return values
based on e.g. end state of the reconciled object.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>