refactor(cli)!: "stage management" sub-commands
This refactors the `kargo stage promote` and
`kargo stage promote-subscribers` commands into a single
`kargo promote` command with `--stage` and
`--subscribers-of` flags.

Effectively, this means for the promotion of a Freight to a Stage, one
can now run:

```shell
kargo promote --project <project> --freight <freight-id> --stage <stage>
```

While for the promotion of a Freight to subscribers of a Stage, one
must run:

```shell
kargo promote --project <project> --freight <freight-id> --subscribers-of <stage>
```

The `kargo stage enable-auto-promotion` and
`kargo disable-auto-promotion` commands remain unchanged, as they will
likely be removed this development cycle.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>