refactor(cli): restructure option flags
This adds more structure to the way option flags are being registered
for a command, while also providing a more standardized way to
validate the option values derived from the flags.

In addition, it ensures flags are registered as required, mutually
exclusive, or taking file and/or directory inputs. Which improves
out-of-the-box validation, while also enriching the shell completion
instructions. For more information around this, see e.g.
https://pkg.go.dev/github.com/spf13/cobra#Command.MarkFlagRequired

By tackling this now, we have a better foundation to more rigorously
start defining a standard structure for commands, splitting out the
`option.Option` configuration, etc.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>