refactor(cli)!: remove `--local-server` flag
As discussed, we do not want to support this on a per-command basis
but would rather (sometime) support spawning a local UI using the user
their Kubeconfig.

In addition, the current flag does not work at all due to the lack of
an auth interceptor injecting user information into the context. Which
cause the local API server to be unable to retrieve a client to
facilitate requests.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>