fix: allow obtaining diff paths for root commit
The previous approach was inherited from when there was a need to be
able to obtain the diff paths for a range of commits. With the
refactoring in v0.7.0, however, this is no longer a requirement.

Because of this, we can switch to `git show` to obtain the diff paths
for a specific commit. Using `--name-only` to print the paths, while
further supplying `--pretty=""` to omit any other commit (message)
data. Effectively yielding the same output as the previous command,
except that it now also works for root commits.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>