fix(api): make protobuf compatible
Resolves issues with protobuf and Kubernetes codegen by adjusting the
data structures. Protobuf has limitations with lists-of-lists, and
Kubernetes codegen does not handle maps with pointer values well.

To fix this, encapsulate the selected Freight within a field and
eliminate the use of pointer map values.

This necessitated the introduction of a helper function to
simplify modifications, particularly within loops. Additionally,
`FreightSelection` is now returned as a pointer from the history
slice.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>