fix(promotion): wait for open pull request
This addresses the immediate issue, by ensuring the Phase of the
Promotion is set to Running for as long as a pull request is open.

However, I do feel this is the result of us not taking a defensive
approach when composing the Phase the Promotion is in, by assuming
a Promotion to be Successful (in multiple places) until we are told
it isn't.

The less error prone approach would be to assume the opposite, but
this is arguably much more difficult to achieve.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>