set region on default config for AWS creds
when falling through the non-project-specific role, the AWS config has a
different region set than when using the project-specific-role.

for the case where the region where kargo is running is different than
the region for the ECR, this causes auth token errors.

fix this by setting the region for the non-project-specific role
codepath.

Signed-off-by: Joel Diaz <jo.diaz@celonis.com>