fix: improve debug information `http` step
This adds the HTTP response code to the error message in case of
failures, and provides a `trace` level log message with the HTTP
response. Combined, this should make it easier to debug a variety
of error scenarios a user may run into when trying to write a `http`
directive.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>