feat!: make `delete` directive non-strict by default
This makes a `delete` step succeed when a file or directory that is
attempted to be removed already does not exist, instead of failing.

In case this behavior is undesired, and the promotion has to fail
when the path does not exist, the `strict` configuration option can
be set to `true`.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>