fix(step): tighten validation `desiredRevision`
This sets the minimum length of the `desiredRevision` field to 40
characters (the lenght of a SHA1 commit reference) to signal to users
that an absolute reference must be configured here (and not e.g., a
branch name).

This will work with OCI in ArgoCD >=3.1 as well, as an image tag is
resolved to a digest in this case (e.g., `sha256:abc123...`).

Signed-off-by: Hidde Beydals <hidde@hhh.computer>