feat(event): Adds a concrete event type
This uses cloudevents to make sure it is easier to support arbitrary
data types and also declares a discrete data type for the currently
existing event.

This is just a Go API change and not a change to actual behavior. The
existing method of publishing as a Kubernetes event is unchanged, but
does give more flexibilty for managing/consuming the event

Signed-off-by: Taylor Thomas <taylor.thomas@akuity.io>