feat(logging)!: Remove logrus and expose logging in `pkg`
This takes some learnings we've had from other projects and folds them
into the logging implementation, as well as making it accessible from
`pkg`. I marked as breaking because we are swapping out the logger, so
logging format might look a little different. Mostly doing this as a heads
up more than out of any particular worry

Signed-off-by: Taylor Thomas <taylor.thomas@akuity.io>