chore: revert `github.com/ktrysmt/go-bitbucket`
The update to `v0.9.88` was paired with a signature change of the
constructor function to properly return the `log.Fatalf` as an error.

We do, however, have upstream dependents with dependencies that have
not updated to this version yet, causing compatibility issues. To
mitigate this, downgrade the dependency again, and work around the
`log.Fatalf` issue by catching the panic.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>